# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SAMT Unified Backend API
  version: 1.0.0
  description: Generated by scripts/generate-openapi.mjs from service OpenAPI docs exposed via API Gateway.
servers:
  - url: http://localhost:9080
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - auth-controller
      summary: Register user
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/identity_RegisterRequest'
        required: true
      responses:
        '201':
          description: User registered
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/identity_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/auth/refresh:
    post:
      tags:
        - auth-controller
      summary: Refresh access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/identity_RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: Refresh succeeded
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/identity_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/auth/logout:
    post:
      tags:
        - auth-controller
      summary: Logout user
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/identity_LogoutRequest'
        required: true
      responses:
        '200':
          description: Logout succeeded
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/identity_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
      security:
        - bearerAuth: []
  /api/auth/login:
    post:
      tags:
        - auth-controller
      summary: Login user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/identity_LoginRequest'
        required: true
      responses:
        '200':
          description: Login succeeded
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/identity_ApiResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '405':
          $ref: '#/components/responses/Error405'
        '409':
          $ref: '#/components/responses/Error409'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  schemas:
...
    identity_ApiResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        message:
          type: string
        path:
          type: string
        correlationId:
          type: string
        degraded:
          type: boolean
...
    identity_RegisterRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 255
          minLength: 0
          format: email
        password:
          type: string
          minLength: 8
          pattern: ^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,128}$
          maxLength: 128
        confirmPassword:
          type: string
          minLength: 1
        fullName:
          type: string
          maxLength: 100
          minLength: 2
          pattern: ^[\p{L}]+(?:[ -][\p{L}]+)*$
        role:
          type: string
          minLength: 1
          pattern: ^STUDENT$
          enum:
            - STUDENT
      required:
        - email
        - password
        - confirmPassword
        - fullName
        - role
    identity_RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
    identity_LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
        - refreshToken
    identity_LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
...
    StandardError:
      type: object
      properties:
        type:
          type: string
          example: https://api.example.com/errors/validation-error
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: Invalid UUID
        instance:
          type: string
          example: /api/users/not-a-number
        timestamp:
          type: string
          format: date-time
...
  responses:
    Error400:
      description: 400 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error401:
      description: 401 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error403:
      description: 403 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error404:
      description: 404 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error405:
      description: 405 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error409:
      description: 409 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error415:
      description: 415 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
    Error500:
      description: 500 error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    identity_bearerAuth:
      type: http
      description: Enter JWT access token
      scheme: bearer
      bearerFormat: JWT
    userGroup_bearerAuth:
      type: http
      description: JWT token from identity-service
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT
    projectConfig_Bearer Authentication:
      type: http
      description: Enter your JWT token from Identity Service
      scheme: bearer
      bearerFormat: JWT
    sync_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
    report_bearerAuth:
      type: http
      description: Paste the JWT only. Swagger UI adds the Bearer prefix automatically.
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT
...